# Improve the performance of ConnState.String()